Recommender System/deployment/requirements.txt
streamlit
pandas
tables
Pillow